Moved treemap series into a subfolder.
