+++ js/masters/modules/treemap.src.js
@@ -9,4 +9,4 @@
  * License: www.highcharts.com/license
  */
 'use strict';
-import '../../Series/TreemapSeries.js';
+import '../../Series/Treemap/TreemapSeries.js';
